msm: mhi_dev: Update ring cache logic
When MHI (modem host interface) device receives a ring start
command do not cache the ring. Cache the ring after a doorbell
for the channel is received. This avoid scenario where host
has added elements to the transfer ring and updated the write
offset as part of start command. Currently device caches the ring
to maintain copies of the write and read offset but in the
process updates the write offset without calling the client
callback indicating there is data. Also fix updating the wrap around
condition when caching the ring.

Change-Id: Iac0aa9fa97c0075cea1431f7e22e3d916532a37a
Signed-off-by: Siddartha Mohanadoss <smohanad@codeaurora.org>