wakeup: Create fake debugfs folder solely for wakeup_sources
edit: adjust for lg debug that seems to be integral to stock rom.
thx to sultanxda for this, I was trying to do this myself earlier
but I don't code :p

Userspace applications and other things expect the wakeup_sources path
to be /sys/kernel/debug/wakeup_sources, not /proc/wakelocks.

To satisfy these userspace components when debugfs is disabled, keep
wakeup_sources in the same location that it's usually expected to be in.

Signed-off-by: Sultanxda <sultanxda@gmail.com>