usb: dwc3: Prevent runtime_suspend after wakeup event
Currently the core is being suspended even after the
wakeup event because the autosuspend delay is elapsed
after the wakeup event is received but before the sm_work
is executed.
Prevent the runtime_suspend being called by marking last_busy
when the wakeup interrupt is received. Also bail out
from the dwc3_msm_suspend if a wakeup interrupt is
received.

Change-Id: I229338adbf3a6719e7e818d1da83eda9f03034b7
Signed-off-by: Sriharsha Allenki <sallenki@codeaurora.org>

Conflicts:
	drivers/usb/dwc3/dwc3-msm.c

LG code.