defconfig: disable lge_display_dyn_dsi_mode_switch
This option facilitates switching between two separate dtsi-files:
dsi-panel-sw49407-dsc-qhd-cmd.dtsi, and
dsi-panel-sw49407-dsc-qhd-cmd-video.dtsi

The first one is patched for engineering aboot, to fix static screen.
The second one is not. I could not do it myself as it is different,
and need different hex values entered.

Disabling this option will fix a problem relating to this on the
relevant devices (H990, F800) causing sceen to not function.