dts: lge: increase bcl mitigation freq (G5,V20)
When battery drops below a certain percentage, any significant
current draw will cause the battery voltage to drop below a set
limit. When this happens bcl will throttle the cpu frequencies.

I'm not sure what the idea is, but having performance drop by
~30% at around half battery capacity probably is undesireable
to most.
Try to ameliorate this by slightly increasing the frequency of
the big cluster during throttling.