sound: msm: Partially remove "Update from H918 20G source"
This extra mixer code seems to cause power drain on V20 as well.
It must not be part of the released stock kernel. The H918 20G kernel
source must be a snapshot from earlier development for this to occur.

Also revert a change in wcd9335.c which I'm unsure about.