kernel: cpu_input_boost: Update accompanying app boosting
kernel: Boost whenever a zygote-forked process becomes a top app

Boost the CPU to the max for 1000 ms whenever the top app changes, which
improves app launch speeds and addresses jitter when switching between
apps. A check to make sure that the top-app's parent is zygote ensures
that a user-facing app is indeed what's added to the top app task group,
since app processes are forked from zygote.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>

------------------------------------------------------------

kernel: Boost to the max for a short amount of time when zygote forks

In order to boost during the period from when a new app is forked to the
time that it is added to the top-app cpuset (which is a signal to
subsequently boost again), boost for a short amount of time when zygote
forks. Not all processes forked from zygote are top apps, hence the
decoupled boosting setup.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>