arm: dts: Add adjustments by LG
* Omit changes to vendor and system dt nodes.

This one had some interesting conflicts. msm8996.dtsi now has a new dtsi file included (msm8996-bus.dtsi) and msm8996-pinctrl had a lot of changes in the gpio layout... could be related to the IR, but it's just a hunch and i am kinda paranoid about that at this point.
In general, the changes are mostly the same as the original commit as well.