msm: mdss: hdmi: override default resolution after parsing EDID
HDMI driver registers fb device with default resolution. This default
resolution is being added to the list of supported modes. But there is
a chance that this default resolution may not be supported by
downstream HDMI sink device. So, always override the default resolution
with first supported mode parsed from EDID.

Change-Id: I5d4ad735f2e0f5bef9582f19e479b0db43623b25
Signed-off-by: Narender Ankam <nankam@codeaurora.org>
Signed-off-by: Ramendra Kumar <ramendra@codeaurora.org>