qcom: idle-v8: ensure msm_pc_debug_counters_phys alignment
msm_pc_debug_counters_phys is a 64-bit variable and hence must be 64-bit aligned.
So far this has happened by accident! Fix the code so this is guaranteed.

This fixes the following build error:

MODPOST vmlinux.o
/tmp/.tmp_vmlinux1.sQjbTK.ltrans59.ltrans.o: in function `msm_cpu_pm_probe':
<artificial>:(.text+0xff78): relocation truncated to fit: R_AARCH64_LDST64_ABS_LO12_NC against symbol `msm_pc_debug_counters_phys' defined in .data section in drivers/soc/qcom/idle-v8.o

warning: one possible cause of this error is that the symbol is being referenced in the indicated code as if it had a larger alignment than was declared where it was defined
<artificial>:(.text+0x1007c): relocation truncated to fit: R_AARCH64_LDST64_ABS_LO12_NC against symbol `msm_pc_debug_counters_phys' defined in .data section in drivers/soc/qcom/idle-v8.o