build: Update and add LTO to build script
* General improvements

* Delete some old cruft

* Add option for building with LTO (GCC)

* Remove option for using ld.gold linker:

I found that (with a spankin new toolchain) gold didn't provide
much benefit in terms of garbage collection, the difference was
~20KB on a kernel of ~12MB. I believe the difference was larger
before, as that must have been the reason I implemented this.

Also since gold seems to be no longer maintained + whispers of
problems with compiling kernel -- I see no reason to keep this.