Make the GPU undervolts actually undervolt based on kernel logs.
So, after looking a bit at some logs provided while testing some
kernels with stendro and xxseva44 i noticed the Adreno GPU
actually sets a voltage that's WAY lower than the ceiling for the
higher freqs, so the undervolt now accounts for that real voltage
and now the GPU can be considered undervolted. Testing it isolated
from the rest of the SoC reports about 15% lower power consumption
or so.