Preparing the power_supply files to work better with anx7688.
The ANX7688 driver treats many parats of the new power_supply_desc
struct as mutable, so it being defined as const in the kernel
power_supply files conflicts with that, so we're making it not
const. Not the safest thing to do, but it should work, and is still
a step up compared to the 3.18 code in general.