Convert anx7688 driver to the new 4.4 format for vars and funcs.
Many changes to write about here, it's been a major overhaul to
most of the code. The only thing worth keeping an eye on is the
dummy_psy variable and some removed checks right below it. The
power_supply_register function changed a bit and now returns a
struct that can't be checked as a boolean to confirm that the power
supply was registered.