Cleanup ANX7688's code and add some extra fixes.
This fixes the misleading message that the chargeer was unknown
and had 0mA as charging current. 3.18 reports the same 0mA but
it waas charging fine there, so this message is incorrect in
regards to the code related to it. The charge problem is caused
by a work scheduling error for now.