Revert a previous queuing method in power_supply_changed.c.
Turns out it borks the improved usb code due to recent changes in dwc3, so we're putting it back to its previous 4.4 schedule call.
Although this fixes a kernel panic, it re-introduces the kernel warning while queuing the anx7688 psy through power_supply_changed. So we need a different approach to fix it.