Revert "drivers: haptics: Port the G6 haptics driver"
This is not the haptics driver, it's an extension to
to it. TouchSense by Immersion Technologies.

"TouchSense technology allows device manufacturers,
content creators and software developers to design
high-quality, power-efficient, tactile effects that
enhance mobile device user interfaces, apps and
entertainment content. Immersion’s TouchSense
software includes haptic software and design
framework, integration tools, reference guides, and
UX design services and support."

This port was not finished, and it needs a userspace
component. If said component is even available it's
possible that it's specifically tuned for the G6 and
wouldn't work properly for the other models.
So just forget about it.

This reverts commit bc023c4947396b9293bdefdf24fdb7d09eacce73.