Cleanup a bit of LG code and improve battery detection in qpnp-fg.
Turns out that instead of importing a lot of lg files for battery
id and etc, we can simply pass the default batt id names defined in
lge_battery_id.h for the fg_batt_type variable and it will get the
correct profile for the devices, enabling battery metrics. Can be
further improved, but for an initial patch, it works.