drivers: power: qcom: Adjust the battery detection code
Commit based on the previous 4.4 branch's code, commit:
690a99300fd5900ca06a9d164d5a0458db42f020.

Instead of relying on LGE usb and psy code, and also
depending on a lot of ifdefs scattered all over the
usb implementations, we're going to pass the default
battery data for each model through a separate file.
The V20 and G5 have removable batteries, but the
available profiles are pretty similar to their
factory-shipped ones, so at most we'll get slightly
inferior charging or power characteristics on better
batteries.

Instead of copying a lot of ifdefs everywhere we're going to
pass battery data , it makes more sense to isolate those to a
special file that handles those things and can be expaned a lot