Revert "Revert "drivers: cpufreq: Add underclock from kerneltoast""
Bring back the underclock!

The reason for this is that, seemingly, the V20 (H990 in this case) can react poorly
to too much peak current drain.
It was reported by a tester; that when his 4G signal was low there was a distinct
possibility of having a kernel crash (that typical lg memdump screen crash ting).
This, according to him, only happened with v0.7, and after eliminating the undervolt
as a culprit, the problem seems to be the combination of higher freqs (draining more
power) and low signal (also causing higher current drain).

This reverts commit 275f2578e2c59de717f9643e44d3f4e85d969495.