arm64: dts: camera: Restore compitable string for jpeg dma
Commit beb54017bb48 changed one compitable string so as to
be picked up by a newer camera2 driver. Problem is, the old
one is still present in the folder, and we've been having
camera issues, so just make both drivers work.

The change was from Sony, so it's not a given that LG's camera
uses the same set of drivers within the camera2 folder.