bluetooth: Bring back LG variant of BRCM V4L2
So, I got bluetooth working with this driver. I was missing a few
patches from Sony Shinano kernel, but they were nowhere to be found
in sonyxperiadev's github, so I almost missed them, and almost gave up.

FM radio isn't working though, but the fm part of the driver seems to
be partly working at least.

I was testing Sony's BRCM V4L2 driver, and it worked simiarly to LG's
but is less stable, at least under the settings currently employed
(there's a lot of them to be tweaked in the fm-deamon/libbt-vendor).
So I think we should go with LG.
Sony abandoned the their driver, and FM support quite early for the
relevant devices. Could be that it's in a slightly broken state.