drivers: usb: misc: tusb422: Disabled usb_psy, and minor fixes
We finally got a boot on the LG G6, marking the first time that
the LGE msm8996 family of devices as a whole booted 4.4. Some usb
changes were needed in tusb422 in order to allow the g6 to boot
isntead of crashing on a power_supply_get_property() call due to
it getting the usb_psy just like anx7688/7418 did previously.