sound: soc: msm: Overhaul CONFIG_SND_DISABLE_DUMMY_DAI.
This is an LGE config, although it's naming is a bit confusing
based off of what it does in msm8996.c. It was also absent from
sound/soc/msm Kconfig, so it was added in there and renamed to
"CONFIG_SND_DISABLE_LGE_DAI_LINKS" as it reflects better what
that config actually changes in the code when enabled.