drivers: leds-qpnp-flash: Make LGE additions be easier to toggle.
Previously the LGE led regulator workarounds relied on
CONFIG_MACH_LGE, a config that's used for many other things and
can't be easily toggled to test specific code, so instead of
dealing that, we'll just rename their instances in qpnp-leds to
something else and make MACH_LGE a single instance at the top to
govern it.

Some debugging calls were also added to the
qpnp_flash_led_prepare_v1() function, as that one was having
issues before.